feat(Demo): stream audio playback from audio element
previously, a we waited for the entire content to download before resolving a promise to then play
the audio blob. this resulted in extensive wait times for some audio files. this new method allows
the audio element to play audio as soon as data buffers, which results in a streaming style
playback.